feat: quase pronto

+++ src/main/java/br/ifrn/edu/jeferson/ecommerce/service/CategoriaService.java
@@ -7,6 +7,7 @@ import br.ifrn.edu.jeferson.ecommerce.exception.BusinessException;
 import br.ifrn.edu.jeferson.ecommerce.exception.ResourceNotFoundException;
 import br.ifrn.edu.jeferson.ecommerce.mapper.CategoriaMapper;
 import br.ifrn.edu.jeferson.ecommerce.repository.CategoriaRepository;
+import br.ifrn.edu.jeferson.ecommerce.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -22,6 +23,8 @@ public class CategoriaService {
     private CategoriaMapper mapper;
     @Autowired
     private CategoriaMapper categoriaMapper;
+    @Autowired
+    private ProdutoRepository produtoRepository;
 
     public CategoriaResponseDTO salvar(CategoriaRequestDTO categoriaDto) {
         var categoria =  mapper.toEntity(categoriaDto);
@@ -43,6 +46,11 @@ public class CategoriaService {
         if (!categoriaRepository.existsById(id)) {
             throw new ResourceNotFoundException("Categoria não encontrada");
         }
+
+        if (produtoRepository.existsByCategorias_Id(id))
+            throw new BusinessException("Não é possível remover a categoria, pois existem produtos associados a ela.");
+
+
         categoriaRepository.deleteById(id);
     }
